+++ src/aioway/__init__.py
@@ -12,11 +12,11 @@ from . import (
     _utils,
     api,
     codecs,
-    dsets,
     emits,
     envs,
     errors,
     indices,
+    io,
     modes,
     relalg,
     sklearn,
