+++ tests/torch/nn/test_iters.py
@@ -7,7 +7,7 @@ from torch import nn
 
 from aioway._iters import TensorIter, sample_mode
 from aioway._utils import AnyDict, is_fake_tensor, torch_fake_mode
-from aioway.dsets import TensorListIter
+from aioway.io import TensorListIter
 from aioway.modes import NnInitThunk
 from aioway.relalg import StackIter
 from aioway.torch.nn import NnUFuncThunk, build_nn_iter
